modified soft limits check
- modified soft limits check to take in account machine geometric transformations
